+++ lib/files/storage/common.php
@@ -141,13 +141,15 @@ abstract class Common implements \OC\Files\Storage\Storage {
 			return false;
 		} else {
 			$directoryHandle = $this->opendir($directory);
-			while (($contents = readdir($directoryHandle)) !== false) {
-				if (!\OC\Files\Filesystem::isIgnoredDir($contents)) {
-					$path = $directory . '/' . $contents;
-					if ($this->is_dir($path)) {
-						$this->deleteAll($path);
-					} else {
-						$this->unlink($path);
+			if(is_resource($directoryHandle)) {
+				while (($contents = readdir($directoryHandle)) !== false) {
+					if (!\OC\Files\Filesystem::isIgnoredDir($contents)) {
+						$path = $directory . '/' . $contents;
+						if ($this->is_dir($path)) {
+							$this->deleteAll($path);
+						} else {
+							$this->unlink($path);
+						}
 					}
 				}
 			}
@@ -224,6 +226,9 @@ abstract class Common implements \OC\Files\Storage\Storage {
 
 	private function addLocalFolder($path, $target) {
 		if ($dh = $this->opendir($path)) {
+			if(!is_resource($dh)) {
+				return  null;
+			}
 			while (($file = readdir($dh)) !== false) {
 				if ($file !== '.' and $file !== '..') {
 					if ($this->is_dir($path . '/' . $file)) {
@@ -241,7 +246,7 @@ abstract class Common implements \OC\Files\Storage\Storage {
 	protected function searchInDir($query, $dir = '') {
 		$files = array();
 		$dh = $this->opendir($dir);
-		if ($dh) {
+		if (is_resource($dh)) {
 			while (($item = readdir($dh)) !== false) {
 				if ($item == '.' || $item == '..') continue;
 				if (strstr(strtolower($item), strtolower($query)) !== false) {
